+++ themes/AP-Default/theme.php
@@ -13,7 +13,8 @@ $host = $app->host;
 	<link rel="stylesheet" href="themes/<?= esc($c['themeSelect']) ?>/style.css">
 	<meta name="description" content="<?= esc($c['description']) ?>">
 	<meta name="keywords" content="<?= esc($c['keywords']) ?>">
-	<script src="//ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+	<script src="js/dist/autosize.js"></script>
+	<script src="js/dist/editInplace.js"></script>
 <?php $app->editTags(); ?>
 </head>
 <body>